Добавляет вывод миниатюр
