+++ js/main.js
@@ -1,1 +1,4 @@
-import './minipic.js';
+import {createPhotos} from './data.js';
+import {renderThumbnails} from './thumbnails.js';
+
+renderThumbnails(createPhotos());
